[enh] Apply changes from pouchDB changes API instead of rebuild FS tree every time

+++ backend/binary.coffee
@@ -248,6 +248,8 @@ module.exports =
 
                 try
                     fs.unlinkSync binaryPath
+                catch e
+                    # nothing
                 finally
                     client.saveFile urlPath, binaryPath, saveBinaryPath
             else
@@ -258,11 +260,15 @@ module.exports =
             if err and err.status isnt 404 then throw new Error err
             else if binaryDoc?
                 pouch.db.remove binaryDoc, ->
-                    if binaryDoc.path? and binaryDoc.path isnt binaryPath
+                    if binaryDoc.path? and binaryDoc.path isnt binaryPath \
+                    and fs.existsSync(binaryDoc.path)
                         fs.renameSync(binaryDoc.path, binaryPath)
                     downloadFile()
             else
                 downloadFile()
 
-        # Check if the binary document exists
-        pouch.db.get doc.binary.file.id, removeBinary
+        if doc.binary?.file?.id?
+            # Check if the binary document exists
+            pouch.db.get doc.binary.file.id, removeBinary
+        else
+            callback null
